Add Detailed Explanation
Added a detailed explanation of the bfs and pagerank methods in the interface.py section of the README. Provided context on why each approach was used, clarified return values, and ensured consistent formatting throughout the section for better readability. Also added information with respect to the Dockerfile and necessary changes in it.